Рефакторинг, readme, клавиатура
