+++ README.md
@@ -1,0 +1,28 @@
+# Проект ToadBot
+
+ToadBot- телеграм-бот, который обладает следующим функционалом:
+1. Присылает рандомную картинку жабки
+2. Умеет выполнять роль калькулятора для 2 чисел
+3. Находит текущее созвездие планет
+4. Поддерживает игру в случайные числа и города
+5. Находит дату ближайшего полнолуния к указанной дате
+6. Считает количество слов в предложении
+
+## Установка
+
+1. Клонируйте репозиторий с github
+2. Создайте виртуальное окружение
+3. Установите зависимости `pip install -r requirements.txt`
+4. Создайте файл `settings.py`
+5. Впишите в `settings.py` переменные:
+```
+API_KEY = "API-ключ бота, выданный телеграм"
+USER_EMOJI = [':frog:',':duck:',':cat:',':smiling_face_with_smiling_eyes:',':winking_face:',':alien_monster:', ':ghost:']
+INITIAL_CITIES = [
+    'Анадырь',
+    'Анапа',
+    'Армавир',
+    ...
+]
+```
+6. Запустите бота командой `python bot.py`
